requirements.txt
python-dotenv
pandas
numpy
pymongo>=4.0
certifi
scikit-learn
mlflow